Provide queue player specific publishers through an associated extension
